fix(board-item): make several changes & update docs
