+++ src/data-display/borad-item/mock.ts
@@ -1,0 +1,37 @@
+import type {
+    IconAction,
+} from '@/data-display/borad-item/type';
+
+export const standardIconActionSet: IconAction[] = [
+    {
+        iconName: 'ic_edit',
+        eventAction: () => {},
+    },
+    {
+        iconName: 'ic_duplicate',
+        eventAction: () => {},
+    },
+    {
+        iconName: 'ic_trashcan',
+        eventAction: () => {},
+    },
+];
+
+export const extraIconActionSet: IconAction[] = [
+    {
+        iconName: 'ic_edit',
+        eventAction: () => {},
+    },
+    {
+        iconName: 'ic_duplicate',
+        eventAction: () => {},
+    },
+    {
+        iconName: 'ic_trashcan',
+        eventAction: () => {},
+    },
+    {
+        iconName: 'ic_search',
+        eventAction: () => {},
+    },
+];
