--- src/data-display/cards/card-list/card-list-item/PCardListItem.stories.mdx
@@ -1,93 +1,0 @@
-import { Meta, Canvas, Story, ArgsTable } from '@storybook/addon-docs/blocks';
-import { i18n } from '@/translations';
-import PCardListItem
-    from "@/data-display/cards/card-list/card-list-item/PCardListItem";
-import { reactive, toRefs, computed } from "vue";
-
-import { getCardListItemArgTypes } from '@/data-display/cards/card-list/card-list-item/story-helper';
-import { getButtonEventHandlerMap } from '@/data-display/cards/card-list/card-list-item/mock';
-
-<Meta title='Data Display/Cards/Card List/Card List Item' parameters={{
-    design: {
-        type: 'figma',
-        url: 'figma url',
-    }
-}} argTypes={getCardListItemArgTypes()}/>
-
-
-export const Template = (args, {argTypes}) => ({
-    props: Object.keys(argTypes),
-    i18n,
-    components: { PCardListItem },
-    template: `
-        <div class="h-full w-full overflow p-8">
-            <p-card-list-item
-                :left-icon="state.leftIcon"
-                :rounded="state.rounded"
-                :right-button-set="state.rightButtonSet"
-                :button-handler-map="state.buttonHandlerMap"
-            >
-                <template #content>
-                    <p>Content</p>
-                </template>
-            </p-card-list-item>
-        </div>
-    `,
-    setup(props) {
-        const state = reactive({
-            leftIcon: computed(() => props.leftIcon),
-            rounded: computed(() => props.rounded),
-            rightButtonSet: computed(() => props.rightButtonSet),
-            buttonHandlerMap: computed(() => props.buttonHandlerMap),
-        });
-        return { state };
-    }
-});
-
-# Card List Item
-<br/>
-<br/>
-
-## Basic
-<br/>
-<br/>
-
-<Canvas>
-    <Story name="Basic">
-        {{
-            i18n,
-            components: { PCardListItem },
-            template: `
-    <div class="h-full w-full overflow p-8">
-        <p-card-list-item
-            left-icon="ic_more"
-            :right-button-set="['EDIT', 'DUPLICATE', 'TRASHCAN']"
-            :button-handler-map="getButtonEventHandlerMap(['EDIT', 'DUPLICATE', 'TRASHCAN'])"
-        >
-            <template #content>
-                <p>Content</p>
-                <p>loooong loooong loooong description.......</p>
-                <p>very very very very loooong loooong loooong description.......</p>
-            </template>
-        </p-card-list-item>
-    </div>
-    `,
-            setup() {
-                return {
-                    getButtonEventHandlerMap,
-                }
-            }
-        }}
-    </Story>
-</Canvas>
-
-
-## Playground
-
-<Canvas>
-    <Story name="Playground">
-        {Template.bind({})}
-    </Story>
-</Canvas>
-
-<ArgsTable story="Playground"/>
