+++ src/index.ts
@@ -36,7 +36,7 @@ export { default as PToggleButton } from './inputs/buttons/toggle-button/PToggle
 export { default as PCard } from './data-display/cards/card/PCard.vue';
 export { default as PCardItem } from './others/deprecated/card-item/PCardItem.vue';
 export { default as PListCard } from './data-display/cards/list-card/PListCard.vue';
-export { default as PCardListItem } from './data-display/cards/card-list/card-list-item/PCardListItem.vue';
+export { default as PBoardItem } from './data-display/borad-item/PBoardItem.vue';
 export { default as PCollapsiblePanel } from './data-display/collapsibles/collapsible-panel/PCollapsiblePanel.vue';
 export { default as PCheckBox } from './inputs/checkbox/PCheckBox.vue';
 export { default as PCheckBoxGroup } from './inputs/checkbox-group/PCheckBoxGroup.vue';
